Add react + express + Node.js
This is the basic of how the server will work. The frontend will be running react to create the user interface. this will be on port 5000. Mostly purpose is for the quiz and practice.
The backend will be running the express. This will help in the database. This will be on port 3000.